requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
mss>=9.0.1
